consolidate conftests
